rename shader -> widget
