+++ src/lib.rs
@@ -1,8 +1,8 @@
 pub mod bitmap;
-pub mod shader;
+pub mod widget;
 
 pub use bitmap::{Bitmap, bitmap};
-pub use shader::surface::{Surface, SurfaceHandler};
-pub use shader::{Controls, TextureCanvas, texture};
+pub use widget::surface::{Surface, SurfaceHandler};
+pub use widget::{Controls, TextureCanvas, texture};
 
-pub use shader::style::{self, Catalog, Status, Style, StyleFn};
+pub use widget::style::{self, Catalog, Status, Style, StyleFn};
